clean up class-fold
